+++ app/src/main/java/com/jlmcdeveloper/notes/ui/main/MainMvpView.java
@@ -4,5 +4,5 @@ import com.jlmcdeveloper.notes.ui.base.MvpView;
 
 public interface MainMvpView extends MvpView {
 
-    void openNoteActivity(int idNode);
+    void openNoteActivity(Long idNode);
 }
