+++ app/src/main/java/com/jlmcdeveloper/notes/util/Constants.java
@@ -1,6 +1,6 @@
 package com.jlmcdeveloper.notes.util;
 
 public class Constants {
-    public final static int ID_NEW_NOTE = -1;
+    public final static Long ID_NEW_NOTE = -1L;
 
 }
